refactor: 인증 관련 클래스 위치 변경
인증 관련 클래스(AuthenticationPrincipal, AuthenticationPrincipalArgumentResolver 등)를 `auth` 패키지에서 `config` 패키지로 이동하여 패키지 구조 정리 및 코드 가독성을 개선했습니다.
