+++ src/main/java/roomescape/config/AuthenticationPrincipal.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.config;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
